use bio

+++ src/http/response.c
@@ -1,38 +1,39 @@
 #define _GNU_SOURCE
 #include "response.h"
+#include "../static.h"
 #include "../utils/assert2.h"
+#include "../utils/compression.h"
 #include "../utils/hash.h"
 #include "../utils/logger.h"
+#include "../utils/static_file.h"
 #include "request.h"
-#include "static.h"
-#include <stdio.h>
-#include <stdlib.h>
 #include <string.h>
 #include <time.h>
 
 char *http_status_code_to_str(http_status_code status_code);
-size_t to_string(const http_response_t *http_response, char *response_str);
-size_t set_body(http_response_t *http_response, const char *uri,
-                char *tmp_buffer);
-size_t set_content_length(http_response_t *http_response,
-                          const size_t content_length);
+unsigned int to_string(const http_response_t *http_response,
+                       char *response_str);
+unsigned int set_body(http_response_t *http_response, const char *uri,
+                      char *tmp_buffer);
+unsigned int set_content_length(http_response_t *http_response,
+                                const unsigned int content_length);
 int set_content_type(http_response_t *http_response, const char *uri);
-size_t set_last_modified(http_response_t *http_response, const char *uri);
-size_t set_date(http_response_t *http_response);
-size_t set_etag(http_response_t *http_response, const char *tmp_body,
-                const size_t tmp_body_size);
-size_t set_compression(http_response_t *http_response,
-                       const char *accept_encoding, char *tmp_buffer);
+unsigned int set_last_modified(http_response_t *http_response, const char *uri);
+unsigned int set_date(http_response_t *http_response);
+unsigned int set_etag(http_response_t *http_response, const char *tmp_body,
+                      const unsigned int tmp_body_size);
+unsigned int set_compression(http_response_t *http_response,
+                             const char *accept_encoding, char *tmp_buffer);
 void handle_if_none_match(http_response_t *http_response,
                           const char *if_none_match);
 
-size_t construct_response(int response_code, const char *uri,
-                          const char *accept_encoding,
-                          const char *if_none_match, char *response,
-                          char *tmp_body_buffer) {
+unsigned int construct_response(int response_code, const char *uri,
+                                const char *accept_encoding,
+                                const char *if_none_match, char *response,
+                                char *tmp_body_buffer) {
   assert(tmp_body_buffer != NULL);
 
-  size_t return_value;
+  unsigned int return_value;
   http_response_t http_response;
 
   // Set status code
@@ -41,7 +42,7 @@ size_t construct_response(int response_code, const char *uri,
   strcpy(http_response.content_language, "en-US");
 
   // Set body
-  size_t content_size = set_body(&http_response, uri, tmp_body_buffer);
+  unsigned int content_size = set_body(&http_response, uri, tmp_body_buffer);
   // Set content length
   set_content_length(&http_response, content_size);
   // Set content type
@@ -57,7 +58,7 @@ size_t construct_response(int response_code, const char *uri,
 
   if (http_response.status_code != HTTP_NOT_MODIFIED) {
     // Set compression
-    size_t compressed_size =
+    unsigned int compressed_size =
         set_compression(&http_response, accept_encoding, tmp_body_buffer);
     // Update content length
     set_content_length(&http_response, compressed_size);
@@ -70,8 +71,8 @@ size_t construct_response(int response_code, const char *uri,
   return return_value;
 }
 
-size_t set_body(http_response_t *http_response, const char *uri,
-                char *tmp_buffer) {
+unsigned int set_body(http_response_t *http_response, const char *uri,
+                      char *tmp_buffer) {
   int return_value = 0;
   if (http_response->status_code != HTTP_OK) {
     const char *status_code_str =
@@ -113,7 +114,8 @@ int set_content_type(http_response_t *http_response, const char *uri) {
   return 0;
 }
 
-size_t set_last_modified(http_response_t *http_response, const char *uri) {
+unsigned int set_last_modified(http_response_t *http_response,
+                               const char *uri) {
   if (http_response->status_code != HTTP_OK) {
     strcpy(http_response->last_modified, "");
     return 0;
@@ -127,15 +129,15 @@ size_t set_last_modified(http_response_t *http_response, const char *uri) {
   }
 }
 
-size_t set_date(http_response_t *http_response) {
+unsigned int set_date(http_response_t *http_response) {
   time_t current_time = time(NULL);
   struct tm *time_info = gmtime(&current_time);
   return strftime(http_response->date, HTTP_HEADER_SMALL_SIZE,
                   "%a, %d %b %Y %H:%M:%S GMT", time_info);
 }
 
-size_t set_etag(http_response_t *http_response, const char *tmp_body,
-                size_t tmp_body_size) {
+unsigned int set_etag(http_response_t *http_response, const char *tmp_body,
+                      unsigned int tmp_body_size) {
   if (http_response->status_code != HTTP_OK ||
       http_response->status_code == HTTP_NOT_MODIFIED) {
     strcpy(http_response->etag, "");
@@ -161,9 +163,9 @@ void handle_if_none_match(http_response_t *http_response,
   }
 }
 
-size_t set_compression(http_response_t *http_response,
-                       const char *accept_encoding, char *tmp_buffer) {
-  size_t return_value = http_response->content_length;
+unsigned int set_compression(http_response_t *http_response,
+                             const char *accept_encoding, char *tmp_buffer) {
+  unsigned int return_value = http_response->content_length;
   int header_count = 0;
   // Validate header
   if (strcasestr(accept_encoding, "gzip") != NULL) {
@@ -195,15 +197,16 @@ size_t set_compression(http_response_t *http_response,
   return return_value;
 }
 
-size_t set_content_length(http_response_t *http_response,
-                          size_t content_length) {
+unsigned int set_content_length(http_response_t *http_response,
+                                unsigned int content_length) {
   http_response->content_length = content_length;
   return content_length;
 }
 
-size_t to_string(const http_response_t *http_response, char *response_str) {
+unsigned int to_string(const http_response_t *http_response,
+                       char *response_str) {
   char *status_code_str = http_status_code_to_str(http_response->status_code);
-  size_t offset = 0;
+  unsigned int offset = 0;
 
   offset += snprintf(response_str + offset, REQUEST_RESPONSE_MAX_SIZE - offset,
                      "HTTP/%s %d %s\r\n", HTTP_VERSION,
